Test porting completed
This patch finishes up the test port for the newly refactored SQL component.
I should mention that we split off the old test file that covered all the
functions in db.go into two new files: one for the persistency interface,
and the other for everything else.

Suggested-by: Maurizio Papini <mpapini@redhat.com>
